python_backend/requirements.txt
fastapi
uvicorn
pydantic
langchain-core
langchain-google-genai
langchain-ollama
ollama
python-dotenv
requests
typing-extensions
